+++ src/molecules/Tag/types.ts
@@ -1,5 +1,6 @@
-import { Tag, GetProps } from 'antd';
+import { GetProps, Tag } from 'antd';
 import { ComponentToken as TagComponentTokenAntd } from 'antd/es/tag/style';
+import { PropsWithChildren, ReactNode } from 'react';
 
 //#region Define Ant Design types
 type TagPropsAntd = GetProps<typeof Tag>;
@@ -13,6 +14,8 @@ type TagComponentTokenExtend = {};
 //#region Define extended types
 type TagPropsExtend = {
     loading?: boolean;
+
+    type?: 'default' | 'text';
 };
 type TagCheckablePropsExtend = {};
 //#endregion
@@ -20,6 +23,12 @@ type TagCheckablePropsExtend = {};
 //#region Export types
 export type RdTagProps = TagPropsAntd & TagPropsExtend;
 export type RdTagCheckableProps = TagCheckablePropsAntd & TagCheckablePropsExtend;
+export interface RdTagGroupProps extends PropsWithChildren {
+    children: ReactNode;
+    maxRows?: number; // Số hàng tối đa, default 1
+    maxCount?: number; // Giới hạn số tag visible (bất kể wrap), optional
+    overflowMode?: 'tooltip' | 'expand' | 'hidden'; // Mode xử lý overflow
+}
 
 export type RdTagComponentToken = TagComponentTokenAntd & TagComponentTokenExtend;
 //#endregion
@@ -28,11 +37,15 @@ export type RdTagComponentToken = TagComponentTokenAntd & TagComponentTokenExten
 export type RdTagComponent = React.ForwardRefExoticComponent<
     RdTagProps & React.RefAttributes<HTMLSpanElement>
 >;
+
+export type RdTagGroupComponent = React.FC<RdTagGroupProps>;
+
 export type RdTagCheckableComponent = React.ForwardRefExoticComponent<
     RdTagCheckableProps & React.RefAttributes<HTMLSpanElement>
 >;
 
 export type RdTagCompoundedComponent = RdTagComponent & {
     CheckableTag: RdTagCheckableComponent;
+    Group: RdTagGroupComponent;
 };
 //#endregion