+++ CHANGELOG.md
@@ -5,11 +5,21 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.11.0]
+
+### Add
+
+-   Add `Filter` component.
+-   Add `Tag.Group` component.
+-   Add `type` prop for `Tag` component.
+-   Add  `useFilter` hooks.
+
 ## [1.10.1]
 
 ### Fixed
 
--   Fix `HeaderSider` in `DashboardTemplate`.
+-   Remove cache when header parameters change for all query.
+-   Hidden learn more wallet in client cPanel
 
 ## [1.10.0]
 
@@ -34,7 +44,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -   Added chart to `SummaryCard` component.
 -   Added `Statistic` component.
 -   Added `Grid.useBreakpoint()` hooks.
--   Added  `sidebarMode` for sidebarProps for `DashboardTemplate`.
+-   Added `sidebarMode` for sidebarProps for `DashboardTemplate`.
 -   Support responsive for `DashboardTemplate`.
 -   Added `LineChart` component.